tasks

+++ 2022.10.10-Homework-4/Project5/Source.cpp
@@ -2,15 +2,15 @@
 
 int main(int argc, char* argv[])
 {
-	int n = 0;
-	double sum = 1.0;
-	int fact = 1;
+	double n = 0;
+	double sum = 1;
+	double fact = 1;
 	std::cin >> n;
 
-	for (int i = 1; i <= n; i++)
+	for (double i = 1; i <= n; i++)
 	{
 		fact = fact * i;
-		sum += 1.0 / fact;
+		sum += 1 / fact;
 	}
 
 	std::cout << sum << std::endl;
